extract request hash lookup route
